+++ src/common.rs
@@ -1,0 +1,3 @@
+pub type Point3 = nalgebra::Vector3<f64>;
+pub type Vec3 = nalgebra::Vector3<f64>;
+pub type Color = nalgebra::Vector3<f64>;