+++ src/hittable.rs
@@ -1,0 +1,13 @@
+use crate::common::*;
+use crate::ray::Ray;
+
+#[derive(Clone, Copy, Debug)]
+pub struct HitRecord {
+    pub point: Point3,
+    pub normal: Vec3,
+    pub t: f64,
+}
+
+pub trait Hittable {
+    fn hit(&self, ray: &Ray, t_min: f64, t_max: f64) -> Option<HitRecord>;
+}