+++ src/ray.rs
@@ -1,15 +1,17 @@
+use crate::common::*;
+
 #[derive(Debug, Copy, Clone)]
 pub struct Ray {
-    pub origin: nalgebra::Vector3<f64>,
-    pub direction: nalgebra::Vector3<f64>,
+    pub origin: Point3,
+    pub direction: Vec3,
 }
 
 impl Ray {
-    pub fn new(origin: nalgebra::Vector3<f64>, direction: nalgebra::Vector3<f64>) -> Self {
+    pub fn new(origin: Point3, direction: Vec3) -> Self {
         Self { origin, direction }
     }
 
-    pub fn at(self, t: f64) -> nalgebra::Vector3<f64> {
+    pub fn at(self, t: f64) -> Point3 {
         self.origin + self.direction * t
     }
 }